cleanup component
